pyproject.toml
[tool.poetry]
name = "jogo_snake_tkinter"
version = "0.1.0"
description = ""
authors = ["Willames <williames@arbache.com>"]

[tool.poetry.dependencies]
python = "^3.9.5"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
